requirements.txt
django==2.0
django-decouple
python-slugify
django-mptt
django-tinymce
Pillow